Rename aidbox.Client to aidbox.ApiClient
