+++ docs/resources/client.md
@@ -1,19 +1,19 @@
 ---
 # generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "aidbox_auth_client Resource - terraform-provider-aidbox"
+page_title: "aidbox_client Resource - terraform-provider-aidbox"
 subcategory: ""
 description: |-
   Client https://docs.aidbox.app/security-and-access-control-1/auth/basic-auth.
 ---
 
-# aidbox_auth_client (Resource)
+# aidbox_client (Resource)
 
 Client https://docs.aidbox.app/security-and-access-control-1/auth/basic-auth.
 
 ## Example Usage
 
 ```terraform
-resource "aidbox_auth_client" "example" {
+resource "aidbox_client" "example" {
   name        = "my-client"
   secret      = "secret"
   grant_types = ["basic"]
