+++ docs/resources/token_introspector.md
@@ -31,10 +31,13 @@ resource "aidbox_token_introspector" "example" {
 ### Optional
 
 - **box_id** (String) ID of box this object lives in
-- **id** (String) The ID of this resource.
 - **introspection_endpoint** (Block List, Max: 1) Configuration for introspecting opaque access tokens. (see [below for nested schema](#nestedblock--introspection_endpoint))
 - **jwks_uri** (String) Location of JWKS public key information for validating JWT tokens
 - **jwt** (Block List, Max: 1) Configuration for validating jwt type access tokens (see [below for nested schema](#nestedblock--jwt))
+
+### Read-Only
+
+- **id** (String) The ID of this resource.
 
 <a id="nestedblock--introspection_endpoint"></a>
 ### Nested Schema for `introspection_endpoint`
