+++ examples/resources/aidbox_client/resource.tf
@@ -1,4 +1,4 @@
-resource "aidbox_auth_client" "example" {
+resource "aidbox_client" "example" {
   name        = "my-client"
   secret      = "secret"
   grant_types = ["basic"]
